BigQuery/requirements.txt
google-cloud-bigquery
python-dotenv
pip install numpy
pip install pandas